jxl-oxide-cli: Refactor video module
